requirements.txt
lark>=1.1.9
google-genai>=0.1.0
pytest>=8.0.0
pytest-html>=4.1.0
PyYAML>=6.0

